+++ qabelbox/src/androidTest/java/de/qabel/qabelbox/ui/views/FileBrowserFragmentTest.kt
@@ -1,27 +1,29 @@
 package de.qabel.qabelbox.ui.views
 
 import android.content.Intent
-import de.qabel.qabelbox.navigation.MainNavigator
-import de.qabel.qabelbox.ui.AbstractUITest
-import de.qabel.qabelbox.ui.idling.InjectedIdlingResource
-import de.qabel.qabelbox.box.presenters.FileBrowserPresenter
-import org.junit.Test
-
 import android.support.test.espresso.Espresso
 import android.support.test.espresso.Espresso.onView
 import android.support.test.espresso.action.ViewActions.click
 import android.support.test.espresso.action.ViewActions.longClick
 import android.support.test.espresso.assertion.ViewAssertions
 import android.support.test.espresso.matcher.ViewMatchers.*
+import android.support.test.espresso.intent.Intents
+import android.support.test.espresso.intent.matcher.IntentMatchers.*
+import org.hamcrest.core.AllOf.allOf
+import com.natpryce.hamkrest.*
 import de.qabel.qabelbox.R
 import de.qabel.qabelbox.activities.MainActivity
-import de.qabel.qabelbox.box.views.FileBrowserFragment
 import de.qabel.qabelbox.box.dto.BrowserEntry
-import de.qabel.qabelbox.helper.UIHelper
-import de.qabel.qabelbox.ui.presenters.ChatPresenter
-import org.hamcrest.core.IsNot
+import de.qabel.qabelbox.box.presenters.FileBrowserPresenter
+import de.qabel.qabelbox.box.views.FileBrowserFragment
+import de.qabel.qabelbox.navigation.MainNavigator
+import de.qabel.qabelbox.ui.AbstractUITest
+import de.qabel.qabelbox.ui.idling.InjectedIdlingResource
+import de.qabel.qabelbox.util.toDownloadSource
 import org.junit.Ignore
-import org.mockito.Mockito.*
+import org.junit.Test
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.verify
 import java.util.*
 
 class FileBrowserFragmentTest: AbstractUITest() {
@@ -29,10 +31,7 @@ class FileBrowserFragmentTest: AbstractUITest() {
 
     lateinit var fragment: FileBrowserFragment
     lateinit var presenter: FileBrowserPresenter
-
-    override fun setUp() {
-        super.setUp()
-    }
+    val file = BrowserEntry.File("Name.txt", 42000, Date())
 
     override fun getDefaultIntent(): Intent {
         return Intent(mContext, MainActivity::class.java).apply {
@@ -64,7 +63,6 @@ class FileBrowserFragmentTest: AbstractUITest() {
     @Test
     fun entryClick() {
         launch()
-        val file = BrowserEntry.File("Name.txt", 42000, Date())
         val entries = listOf(file)
         fragment.showEntries(entries)
 
@@ -75,7 +73,6 @@ class FileBrowserFragmentTest: AbstractUITest() {
     @Test
     fun bottomSheetForFiles() {
         launch()
-        val file = BrowserEntry.File("Name.txt", 42000, Date())
         fragment.showEntries(listOf(file))
         onView(withText(file.name)).perform(longClick())
         listOf(R.string.Send, R.string.Delete, R.string.Export).forEach {
@@ -92,5 +89,20 @@ class FileBrowserFragmentTest: AbstractUITest() {
         listOf(R.string.Delete).forEach {
             onView(withText(R.string.Delete)).check(ViewAssertions.matches(isDisplayed()))
         }
+    }
+
+    @Ignore
+    @Test
+    fun openFile() {
+        launch()
+        fragment.open(file, "content".toDownloadSource())
+        Intents.intended(allOf(hasAction(Intent.ACTION_CHOOSER),
+                hasExtra(Intent.EXTRA_INTENT,
+                        allOf(hasAction(Intent.ACTION_SEND)))))
+    }
+
+    @Test
+    fun uploadFile() {
+        launch()
     }
 }
